requirements.txt
# Python Dependencies for Visual Resume Dashboard ETL Pipeline
# Install with: pip install -r requirements.txt

# Database & ORM
psycopg2-binary==2.9.10
SQLAlchemy==2.0.27

# Data Processing
pandas==2.2.0
numpy==1.26.4

# Excel I/O
openpyxl==3.1.5

# Google Sheets API
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.104.0

# Utilities
python-dotenv==1.0.0
